add missing folder
When building the container on Ubunut 20.04, installation of the Omada deb
package failed. It is assumed that a change in mkosi (version 5) or debootstrap
(version 1.0.118ubuntu1.2) triggered a change in the container image that is
built before mkosi.postinst is run. One of the Debain packages reqiured by the
Omada Controller apparently assums a folder structure exists, which does not.

Until Debain fixes the problem (looks like other people have run into the
problem too), we will create the folder before installing the Omada Controller.